Encapsulate various helper functions
